+++ src/app/manage_course_videos/course_video.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { VideoData } from "../../types";
 import { useAppDispatch, useAppSelector } from "../store/store";
 import * as customActions from "../store/custom_actions";
-import * as asyncActions from "../use_remove_course";
+import * as asyncActions from "../async_actions";
 import { selectVideoStatus, FullUserVideo } from "../video_selectors";
 import { formatBytes } from "../utils/format_bytes";
